Moved AppCenter code to MainViewModel
